Implement change request
Refactored the display of Tic-Tac-Toe in display_banner method
Removed gets.chomp and puts from every file and placed them inside main.rb
Redesigned the valid_name method to fix the bug of accepting empty spaces
