+++ lib/game_board.rb
@@ -6,14 +6,6 @@ class Board
     @board = [1, 2, 3, 4, 5, 6, 7, 8, 9]
   end
 
-  def draw
-    puts "|#{@board[0]} | #{@board[1]} | #{@board[2]}|"
-    puts '---+--------'
-    puts "|#{@board[3]} | #{@board[4]} | #{@board[5]}|"
-    puts '---+--------'
-    puts "|#{@board[6]} | #{@board[7]} | #{@board[8]}|"
-  end
-
   def valid_move?(position)
     position.to_i.between?(1, 9) && !position_taken?(position)
   end
